requirements.txt
langchain 
langchain-openai 
langchain-community
openai 
pinecone-client 
python-dotenv
docx2txt
chromadb
pypdf 
pyyaml
langchain-google-genai
underthesea
sentence-transformers
pywin32
langgraph   
wikipedia
docling 
langchain-docling    